refactor: 소켓 클라이언트 파일명 변경 및 client 계층이 아닌 service 계층 활용하도록 변경
