+++ apps/frontend/src/feature/room/service/room.ts
@@ -1,4 +1,4 @@
-import { SocketClient } from '@/shared/socket/socket';
+import { SocketClient } from '@/shared/socket/client';
 import type { JoinRoomRequest, ServerToClientEvents } from '@plum/shared-interfaces';
 
 /**
