+++ apps/frontend/src/shared/socket/client.ts
